app.js opgeschoond met 1 elememnt
